feat(freetype): 调整SizeMetrics的YPPEM返回类型为Int
- 将`sizeMetricsGetYPPEM`返回类型从`Long`调整为`Int`，统一类型定义
- 更新相关Kotlin、C++以及跨平台接口，实现类型兼容性修改
- 优化代码规范，以提高数据精度与一致性

+++ katex-core/src/androidMain/kotlin/com/pvporbit/freetype/FreeTypeAndroid.kt
@@ -188,7 +188,7 @@ object FreeTypeAndroid : IFreeType {
 
     external override fun sizeMetricsGetXScale(sizeMetrics: Long): Long
 
-    external override fun sizeMetricsGetYPPEM(sizeMetrics: Long): Long
+    external override fun sizeMetricsGetYPPEM(sizeMetrics: Long): Int
 
     external override fun sizeMetricsGetYScale(sizeMetrics: Long): Long
